+++ src/mongo/db/modules/enterprise/jstests/fle2/basic_replace.js
@@ -9,6 +9,12 @@ load("jstests/fle2/libs/encrypted_client_util.js");
 (function() {
 'use strict';
 
+// TODO: SERVER-72932 remove when v2 update is implemented
+if (isFLE2ProtocolVersion2Enabled()) {
+    jsTest.log("Test skipped because featureFlagFLE2ProtocolVersion2 is enabled");
+    return;
+}
+
 let dbName = 'basic_update';
 let dbTest = db.getSiblingDB(dbName);
 dbTest.dropDatabase();
