+++ src/mongo/db/modules/enterprise/jstests/fle2/find_and_modify_replace.js
@@ -8,6 +8,12 @@ load("jstests/fle2/libs/encrypted_client_util.js");
 (function() {
 'use strict';
 
+// TODO: SERVER-72933 remove when v2 findAndModify works
+if (isFLE2ProtocolVersion2Enabled()) {
+    jsTest.log("Test skipped because featureFlagFLE2ProtocolVersion2 is enabled");
+    return;
+}
+
 const dbName = 'find_and_modify_replace';
 const dbTest = db.getSiblingDB(dbName);
 dbTest.dropDatabase();
