+++ src/mongo/db/modules/enterprise/jstests/fle2/query/find_and_modify.js
@@ -14,6 +14,12 @@
 
 load("jstests/fle2/libs/encrypted_client_util.js");
 
+// TODO: SERVER-72933 remove when v2 findAndModify works
+if (isFLE2ProtocolVersion2Enabled()) {
+    jsTest.log("Test skipped because featureFlagFLE2ProtocolVersion2 is enabled");
+    return;
+}
+
 const dbName = 'query_find_and_modify';
 const dbTest = db.getSiblingDB(dbName);
 dbTest.dropDatabase();
