+++ src/mongo/db/modules/enterprise/jstests/fle2/query/range/one_sided_find.js
@@ -20,6 +20,12 @@ if (!isFLE2RangeEnabled()) {
     return;
 }
 
+// TODO: SERVER-72926 remove when v2 find works
+if (isFLE2ProtocolVersion2Enabled()) {
+    jsTest.log("Test skipped because featureFlagFLE2ProtocolVersion2 is enabled");
+    return;
+}
+
 let dbName = jsTestName();
 let dbTest = db.getSiblingDB(dbName);
 dbTest.dropDatabase();
