+++ src/mongo/db/modules/enterprise/jstests/fle2/txn_find_and_modify.js
@@ -12,6 +12,12 @@ load("jstests/fle2/libs/encrypted_client_util.js");
 (function() {
 'use strict';
 
+// TODO: SERVER-72933 remove when v2 findAndModify works
+if (isFLE2ProtocolVersion2Enabled()) {
+    jsTest.log("Test skipped because featureFlagFLE2ProtocolVersion2 is enabled");
+    return;
+}
+
 let dbName = 'txn_find_and_modify';
 let dbTest = db.getSiblingDB(dbName);
 dbTest.dropDatabase();
