+++ src/mongo/db/modules/enterprise/jstests/fle2/txn_update_range.js
@@ -18,6 +18,12 @@ if (!isFLE2RangeEnabled()) {
     return;
 }
 
+// TODO: SERVER-72932 remove when v2 update is implemented
+if (isFLE2ProtocolVersion2Enabled()) {
+    jsTest.log("Test skipped because featureFlagFLE2ProtocolVersion2 is enabled");
+    return;
+}
+
 let dbName = 'txn_insert_range';
 let dbTest = db.getSiblingDB(dbName);
 dbTest.dropDatabase();
